Pretty significant changes have been accomplished here, though the commit itself is small.
DbOobjectsTest: Removed attribute test. That would require screen scraping,
and that is expressly forbidden by geocaching.com. Attributes will have to
be removed entirely, not that much was there before.

Added a simple log uploader gui. It allows the user to specify the accounts
that will receive the logs that have been entered.

Changed the logdate in the main window to a DatePickerCtrl. Made more sense
for what it was.

Forced log list to reload after saving log entries, since the date could
change.

Added code which should upload a log to geocaching.com. Code is not yet
tested, though it is believed it should work just fine.

Modified gpxSyncAll to loop over the accounts/logs. It doesn't upload the
logs yet (nor even attempt to), but this should be very quick to
accomplish.

gpxsource.py: Added custom grid table class for the log uploader window.
Added constructor/etc to allow the uploader window to use the grid table
class correctly.